No need to cast to double, already doubles now

+++ MATLAB/functions/fap.m
@@ -1,6 +1,6 @@
 function x = fap(img, filter)
 %FAP applies 2 given X and Y filters to an image given as input,
 %calculates the magnitude of change from the 2
-x = conv2(double(img), double(filter), 'same');
+x = conv2(img, filter, 'same');
 
 end
